Modify the code model to support annotations
This commit removes the OdcmAnnotated object and modifies the remainder of
objects in the code model so that vocabulary annotations can be
represented
in the code model.

Almost all entities in an Edm should be annotatable by at least a
Core.Description.
As such, almost anything that is an object should have a collection of
annotations

OData defines annotations in a way that most closely represents a named
triple of Namespace, Name, and Value. As such, the code dom representation
of an object should be a string, string and an object, Consumers
of the annotation value can accordingly reference the value for
annotations they know of by casting the object.