Add representations of ComplexTypes in Capabilities vocab
There are a few different complex types that are used in the
OneNote metadata document. This commit adds those complex types so that
they may be parsed and represented.

This is obviously incomplete coverage of the annotations. Additionally,
this code should not be hand written and probably should be generated.