Add the ODataVocabularyReader
Vocabulary annotations are not exposed with the niceset of APi's at the
OData level. This commit adds a utility class to hide the interpretation
of annotation values from EdmModels and returns instances of
OdcmVocabularyAnnotation.

This code leverages reflection to create instances of annotation complex
types
and mostly uses the EdmLib's EdmExpresion evaluation facilities, except
in some small cases.

In order to properly interpret the vocabulary, an xml resource is added
which has the latest version of the capabilities vocabulary from OASIS.