CSharpWriter Collection Count Feature
Adds the ability to call CountAsync on collections generated by the CSharpWriter.
Uses our new Features concept to contain the code artifacts produced by this feature.
When we enable Count Restriction annotations, taking advantage of them will be as easy as
disabling this feature for Entities affected by the annotation.