Support for Capability Annotations Reading in OdcmModel
We are starting with a limited set of Capabilities - InsertRestriction, UpdateRestriction, DeleteRestriction and ExpandRestriction. Capabilities define how an EntityType or its collection is projected in the generated code. To enable that OdcmModel introduces a new class called OdcmProjection. This class contains the OdcmType (EntityType to be projected) and the capabilities (the operations to be projected for the EntityType). On the other hand each OdcmType will maintain a list of corresponding OdcmProjections. The OdcmWriters will use this projection information to drive the codegen of interfaces (or similar artifacts) for a given OdcmType. The emitted interfaces will only project those capabilities that are available when accessing an EntityType (or its collection) from a given context.

In the OdcmModel, both Entity Sets (of an EntityContainer) and Navigation Properties (of an EntityType) are both modeled into OdcmProperty. Given that the Capabilities we are supporting will affect Client Library properties which give access to Entities and Entity sets, the OdcmProperty class will be updated to expose a Projection property. OdcmProjection is added as a property to the OdcmProperty class, so each OdcmProperty is mapped to an OdcmProjection.  [BreakingChange] Also the OdcmType property of the OdcmProperty class is now redundant and will be removed [/BreakingChange].

Essentially there are three major changes in the OdcmModel.
- OdcmType will have a list of OdcmProjections. The OdcmWriter will generate interfaces for each Projection of an OdcmType.
- OdcmProperty will have a reference to an OdcmProjection. OdcmWriter will resolve this Projection to one of the interfaces generated in the above step. That resolved interface will be used as the return type of the Property when codegening.
- OdcmProperty class will no longer have OdcmType property.