Removal of Capabilities Annotations parsing logic from the OdcmModel
OdcmModel is designed to be protocol agnostic. Earlier OData concepts like Capabilities annotations parsing were wrongly introduced into the OdcmModel in the spirit of code simplification. But it violated a core principle of the OdcmModel that it should be protocol agnostic. The fix is to abstract away OData Capabilities annotations parsing into the ODataReader.

Fix #60