Remove the parsing of annotations by 'ODataVocabularyReader'.
ODataVocabularyReader does not gracefully fail for unknown annotations. It aggressively throws exceptions whenever it encounters annotations it is unable to parse. "ODataVocabularyReader.GetOdcmAnnotations" parses and creates a list of 'OdcmVocabularyAnnotation'. But this list not currently used by the odcmwriters to do anything useful. So I have removed the call to "ODataVocabularyReader.GetOdcmAnnotations" as a workaround. We can add this call to parse annotations once it is stable enough to handle unknown annotations. See issue #84.