Add methods AddLinkAsync/RemoveLinkAsync/UpdateAsync/DeleteAsync to the Collection Fetcher Classes
- AddLinkAsync - Adds an entity reference to an entity collection (entity sets, collection-valued navigation properties)
- RemoveLinkAsync - Removes an entity reference from an entity collection (entity sets, collection-valued navigation properties)
- UpdateAsync - Update a single entity in an entity collection (entity sets, collection-valued navigation properties)
- DeleteAsync - Delete a single entity from an entity collection (entity sets, collection-valued navigation properties)

The data-classes from now will only have properties that hold an entity value. It does not send/receive any request/response to/from an OData service. Any operation requiring to talk to the OData service has to go through the Fetcher and Collection classes.

The data classes still explicitly implement the data-interface navigation properties. Ideally this is not required, just the explicit properties are sufficient. But due to a bug in 'Expand' scenario this is not yet removed in the generated proxies.